Charge realyer fee (#128)
* Improve relayer script

* Remove unused relayer code

* Charge relayer fee

  * Add relay event

* Get aptos relay event

* Get eth relay events

* Use pool vaa nonce for eth and aptos

* Fix eth compile

* Test eth relayer

* Fix relayer for aptos

* Fix decimal

* Fix sui portal liquidity and relay event

* Consider relay fee for sui relayer

* refactor: Refactor and optimize relayer script.

- Refactored code for easier readability and maintainability
- Removed unnecessary imports and modified variable types for more efficient code
- Added a lock to prevent race condition issues in `BridgeDict` class
- Improved error messages for successful transactions
- Changed the way to retrieve `source_chain` for better accuracy
- Formatted long logging messages for better readability

* refactor: Refactor SUI relayer and lending SDK with optimizations

- Add `claim_test_coin` function to `dola_sui_sdk.init`
- Modify `sui_portal_watcher` to query relay events with a limit of 5 and use moveEvent fields in variable assignments
- Add logging to identify identified transactions in `sui_portal_watcher`
- Add check to ensure relay_fee_record exists before processing transaction in `sui_pool_executor`
- Use `wormhole_adapter_pool` instead of `bridge_pool` in `sui_pool_executor` variable assignments
- Remove unused imports and commented out code in affected files

* refactor: Refactor pool executor functions to improve relay fee handling.

- Update relayer.py script to handle missing keys in relay_fee_record dictionary
- Re-add message to queue and await for 1 second if key is missing
- Modify sui_pool_executor to include sui_omnipool.dola_pool.PoolApproval[-1] parameter
- Modify aptos_pool_executor to include ty_args=[token_name] in wormhole_adapter_pool.receive_withdraw
- Modify eth_pool_executor to include ethereum_account parameter in ethereum_wormhole_bridge.receiveWithdraw with value from dola_ethereum_sdk.get_account()

* Split the relayer into chains to execute

* Fix bscscan rpc

* refactor: Refactor relayer and sdk functions.

- Refactor `eth_portal_watcher` to use `start_block` and `end_block` parameters in `lending_relay_event`
- Update code to store `relay_fee_record` instead of modifying `data`
- Rename `build_bsc_rpc_params` to `build_rpc_params` and remove `build_polygon_rpc_params`
- Decrease the amount monitored for `usdt()` supply by a factor of 10

* refactor: Refactor lending functionality and improve relayer retries.

- Update lending parameters for withdrawal monitoring in Ethereum SDK
- Modify relayer functions for better performance and error handling in SUI SDK
- Comment out and uncomment certain function calls in both SDKs as needed

* Add basic compensation func

* refactor: Refactor and optimize contract asset transfer and function naming conventions

- Implement `transferAsset` for `relayer.transfer` calls in `System.sol`
- Refactor `binding` and `unbinding` functions to reduce code duplication in `System.sol`
- Replace `relayer.transfer` with `LibAsset.transferAsset` in `Lending.sol`
- Refactor `supply` and `repay` functions in `Lending.sol` to use `LibAsset.queryDecimals`
- Refactor `amount` in several functions using `LibDecimals.fixAmountDecimals` in `Lending.sol`
- Rename `liquidate` function to `liquidateCollateral` in `Lending.sol`
- Add `uint16` type definition to several parameters in `Lending.sol`

* refactor: Refactor relay event functions and combine data from multiple portals

- Refactor multiple functions in the ethereum and aptos SDKs for improved functionality
- Create a new function in the aptos SDK to combine previously modified functions
- Move portals code into individual functions and fix code typo in relayer.py script
- Use correct fee amount in eth_portal_watcher script.

+++ ethereum/contracts/omnipool/WormholeAdapterPool.sol
@@ -28,8 +28,8 @@ contract WormholeAdapterPool {
     mapping(bytes32 => bool) consumedVaas;
 
     // todo! Delete after wormhole running
-    mapping(uint32 => bytes) public cachedVAA;
-    uint32 vaaNonce;
+    mapping(uint64 => bytes) public cachedVAA;
+    uint64 vaaNonce;
 
     event PoolWithdrawEvent(
         uint64 nonce,
@@ -52,7 +52,7 @@ contract WormholeAdapterPool {
     }
 
     // todo! Delete after wormhole running
-    function getNonce() public view returns (uint32) {
+    function getNonce() public view returns (uint64) {
         return vaaNonce;
     }
 
@@ -79,7 +79,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_REGISTER_OWNER,
             "INVALID CALL TYPE"
@@ -97,7 +97,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_DELETE_OWNER,
             "INVALID CALL TYPE"
@@ -115,7 +115,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_REGISTER_SPENDER,
             "INVALID CALL TYPE"
@@ -133,7 +133,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_DELETE_SPENDER,
             "INVALID CALL TYPE"
@@ -159,13 +159,13 @@ contract WormholeAdapterPool {
             LibAsset.maxApproveERC20(IERC20(token), address(dolaPool), amount);
         }
 
-        bytes memory payload = dolaPool.deposit{value: msg.value - wormholeFee}(
+        bytes memory payload = dolaPool.deposit{value : msg.value - wormholeFee}(
             token,
             amount,
             appId,
             appPayload
         );
-        wormhole.publishMessage{value: wormholeFee}(
+        wormhole.publishMessage{value : wormholeFee}(
             0,
             payload,
             wormholeFinality
@@ -177,13 +177,13 @@ contract WormholeAdapterPool {
 
     /// Send message that do not involve incoming or outgoing funds by application
     function sendMessage(uint16 appId, bytes memory appPayload)
-        external
-        payable
+    external
+    payable
     {
         uint256 wormholeFee = wormhole.messageFee();
         require(msg.value >= wormholeFee, "FEE NOT ENOUGH");
         bytes memory payload = dolaPool.sendMessage(appId, appPayload);
-        wormhole.publishMessage{value: msg.value}(0, payload, wormholeFinality);
+        wormhole.publishMessage{value : msg.value}(0, payload, wormholeFinality);
         cachedVAA[getNonce()] = payload;
         increaseNonce();
     }
@@ -198,7 +198,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.WithdrawPayload memory payload = LibPoolCodec
-            .decodeWithdrawPayload(encodedVm);
+        .decodeWithdrawPayload(encodedVm);
         dolaPool.withdraw(payload.user, payload.amount, payload.pool);
 
         emit PoolWithdrawEvent(
